requirements.txt.txt
opencv-python
mediapipe
numpy
pygame
Pillow
